4차 머지

+++ backend/src/main/java/com/example/Piroin/project/domain/deposit/repository/DepositRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface DepositRepository extends JpaRepository<Deposit, Long> {
+public interface DepositRepository extends JpaRepository<Deposit, Integer> {
     Optional<Deposit> findByUser(User user);
 
     Optional<Deposit> findByUserId(Long userId);
